restructure: move project files into the appropriate directories, created makefile for print-test project
