+++ src/projects/STM32G431/B-G431B-ESC1/main/STM32G431KBTX_FLASH.ld
