+++ src/projects/STM32G431/B-G431B-ESC1/main/main.c
