+++ src/projects/STM32G431/B-G431B-ESC1/print_test/printing-test.c
@@ -1,4 +1,4 @@
-#include "G431/platform_g431.h"
+#include "platform_g431.h"
 
 #include "printing.h"
 #include <float.h>
