+++ src/projects/STM32G431/platform_g431.h
