+++ src/projects/STM32G431/startup_stm32g431kbtx.s
